move notFoundLink and serverSideErrorLink directly to components
